Finally the pulses are going into the ROOT file

+++ PSD/include/PSD_PMT_Hit.h
@@ -2,15 +2,14 @@
 #define PSD_PMT_HIT_HH
 
 #include "G4VHit.hh"
-
 class PSD_PMT_Hit : public G4VHit {
-  unsigned long long fPhotonArrivalTime;
+  int fPhotonArrivalTime;
 
 public:
   PSD_PMT_Hit();
   virtual ~PSD_PMT_Hit();
-  void SetArrivalTime(unsigned long long tme);
-  unsigned long long GetArrivalTime() const;
+  void SetArrivalTime(int tme);
+  int GetArrivalTime() const;
   void Print();
 };
